refactor(commands): start the commands/ submodule — extract the pipeline/recipes/saved-views domain
Pure MOVE, zero behavior change. First of the God-file split: commands.rs
(37k lines) becomes commands/mod.rs (git rename) and its
pipeline/recipes/saved-views domain — 8 #[tauri::command] fns
(list/save/delete recipe, list/upsert/delete/reorder saved_view) +
is_valid_saved_view_scope + the scope consts — moves verbatim into
commands/pipeline.rs, re-exported via 'pub use' so every
generate_handler![commands::save_recipe] path resolves unchanged (lib.rs
untouched). A no-gate/no-seal domain (saved views store FE view metadata,
not content).

Establishes the reusable pattern for the remaining domain extractions:
rename-to-mod.rs + per-domain file + glob re-export = zero lib.rs churn.
Trap baked in: a submodule whose name shadows a crate-level 'use
crate::{X}' needs '#[path] mod X_commands; pub use X_commands::*;' to
avoid an E0255 name collision (pipeline hit it).

Tests stay in mod.rs (shared fixtures); test count 2007 before and after.
Verified: cargo test --lib saved_view/recipe green, clippy --lib
-D warnings clean, MSRV grep clean, lib.rs diff empty.
